+++ Sources/KeyboardKit/_Deprecated/EnglishSecondaryCalloutActionProvider.swift
@@ -1,4 +1,0 @@
-import Foundation
-
-@available(*, deprecated, renamed: "EnglishCalloutActionProvider")
-public typealias EnglishSecondaryCalloutActionProvider = EnglishCalloutActionProvider