+++ Sources/KeyboardKit/_Deprecated/KeyboardActionRow.swift
@@ -1,4 +1,0 @@
-import Foundation
-
-@available(*, deprecated, message: "Use KeyboardActions instead")
-public typealias KeyboardActionRow = KeyboardActions