+++ Sources/KeyboardKit/_Deprecated/KeyboardAppearance+Deprecated.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-public extension KeyboardAppearance {
-    
-    @available(*, deprecated, renamed: "actionCalloutStyle")
-    func secondaryInputCalloutStyle() -> ActionCalloutStyle {
-        actionCalloutStyle()
-    }
-}