+++ Sources/KeyboardKit/_Deprecated/KeyboardLocale+Deprecated.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-public extension KeyboardLocale {
-    
-    @available(*, deprecated, renamed: "portuguese_brazil" )
-    static var brazilian: KeyboardLocale { .portuguese_brazil }
-}