+++ Sources/KeyboardKit/_Deprecated/RowItem.swift
@@ -1,4 +1,0 @@
-import Foundation
-
-@available(*, deprecated, renamed: "KeyboardRowItem")
-public typealias RowItem = KeyboardRowItem