+++ Sources/KeyboardKit/_Deprecated/SecondaryInputCallout.swift
@@ -1,4 +1,0 @@
-import Foundation
-
-@available(*, deprecated, renamed: "ActionCallout")
-public typealias SecondaryInputCallout = ActionCallout