+++ Sources/KeyboardKit/_Deprecated/SecondaryInputCalloutContext.swift
@@ -1,4 +1,0 @@
-import Foundation
-
-@available(*, deprecated, renamed: "ActionCalloutContext")
-public typealias SecondaryInputCalloutContext = ActionCalloutContext