+++ Sources/KeyboardKit/_Deprecated/SecondaryInputCalloutStyle.swift
@@ -1,4 +1,0 @@
-import Foundation
-
-@available(*, deprecated, renamed: "ActionCalloutStyle")
-public typealias SecondaryInputCalloutStyle = ActionCalloutStyle