+++ Sources/KeyboardKit/_Deprecated/StandardSecondaryCalloutActionProvider.swift
@@ -1,4 +1,0 @@
-import Foundation
-
-@available(*, deprecated, renamed: "StandardCalloutActionProvider")
-public typealias StandardSecondaryCalloutActionProvider = StandardCalloutActionProvider