+++ Sources/KeyboardKit/_Deprecated/SystemAudioPlayer+Deprecated.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-public extension SystemAudioPlayer {
-    
-    @available(*, deprecated, renamed: "play")
-    func playSystemAudio(_ audio: SystemAudio) {
-        play(audio)
-    }
-}